formatting

+++ src/server.c
@@ -144,8 +144,7 @@ task_t *monitor_run_task (task_t *task)
   }
 
   /* Without these ignore when, for example, ctrl-c is pressed in terminal
-   * all child exec processes are terminated.
-   */
+   * all child exec processes are terminated. */
   signal (SIGINT, SIG_IGN);
   signal (SIGTERM, SIG_IGN);
 
@@ -186,7 +185,7 @@ task_t *monitor_run_task (task_t *task)
   cclose (client_fd);
 
   char buf[128];
-  const int nbytes = snprintf (buf,128,"%c%ld",FINISHED_TASK,(long)getpid());
+  const int nbytes = snprintf (buf, 128, "%c%ld", FINISHED_TASK, (long) getpid ());
   wwrite (g.server_fifo_wr, buf, nbytes + 1);
   fprintf (stderr, "[%ld] monitor exiting\n", (long) getpid ());
 
@@ -242,7 +241,6 @@ void decrement_active_transformations_count (const task_t *task)
     g.get_transformation_active_count[j] -= task->ops_totals[j];
 }
 
-
 /*!
  * task_message ::= ⟨proc_file⟩ | ⟨status⟩ | ⟨finished_task⟩
  *      ⟨proc_file⟩ ::= ⟨client_pid_str⟩ ⟨PROC_FILE⟩ ⟨priority⟩ ⟨src⟩ ⟨dst⟩ ⟨num_ops⟩ ⟨ops⟩⁺
@@ -257,22 +255,23 @@ void process_message (char *m)
 {
   fprintf (stderr, "[%ld] process_message: %s\n", (long) getpid (), m);
 
-  if (m[0] == FINISHED_TASK) {
-    fprintf (stderr, "[%ld] process_message: finished_task\n", (long) getpid ());
-    pid_t monitor_pid = sstrtol (&m[1]);
-    // search for task that was being executed by the monitor that just finished
-    int finished_task_index;
-    for (finished_task_index = 0; finished_task_index < GLOBAL_MAX_PARALLEL_TASKS; ++finished_task_index)
-      if (g.active_tasks[finished_task_index] != NULL
-          && g.active_tasks[finished_task_index]->monitor == monitor_pid)
-        break;
-    assert (finished_task_index < GLOBAL_MAX_PARALLEL_TASKS);
-    decrement_active_transformations_count (g.active_tasks[finished_task_index]);
-    free_task (g.active_tasks[finished_task_index]);
-    g.active_tasks[finished_task_index] = NULL;
-    --g.num_active_tasks;
-    return;
-  }
+  if (m[0] == FINISHED_TASK)
+    {
+      fprintf (stderr, "[%ld] process_message: finished_task\n", (long) getpid ());
+      pid_t monitor_pid = sstrtol (&m[1]);
+      // search for task that was being executed by the monitor that just finished
+      int finished_task_index;
+      for (finished_task_index = 0; finished_task_index < GLOBAL_MAX_PARALLEL_TASKS; ++finished_task_index)
+        if (g.active_tasks[finished_task_index] != NULL
+            && g.active_tasks[finished_task_index]->monitor == monitor_pid)
+          break;
+      assert (finished_task_index < GLOBAL_MAX_PARALLEL_TASKS);
+      decrement_active_transformations_count (g.active_tasks[finished_task_index]);
+      free_task (g.active_tasks[finished_task_index]);
+      g.active_tasks[finished_task_index] = NULL;
+      --g.num_active_tasks;
+      return;
+    }
 
   const char del[2] = "\31";
   char *tok;
@@ -360,8 +359,10 @@ void process_message (char *m)
 }
 
 void sig_handler (__attribute__((unused)) int signum)
-{ unlink (SERVER);
-  g.has_been_interrupted = 1; }
+{
+  unlink (SERVER);
+  g.has_been_interrupted = 1;
+}
 
 void block_read_fifo ()
 {
@@ -374,7 +375,7 @@ void block_read_fifo ()
   if (g.has_been_interrupted)
     return;
   assert(nbytes > 0 && nbytes < BUFSIZ);
-  fprintf (stderr, "[%ld] block_read: read(s=%zu) %.*s\n", (long) getpid (), nbytes, nbytes, buf);
+  fprintf (stderr, "[%ld] block_read: read(s=%zu) %.*s\n", (long) getpid (), nbytes, (unsigned int) nbytes, buf);
   for (size_t i = 0; i < nbytes;)
     {
       size_t j = strlen (&buf[i]) + 1;
@@ -426,8 +427,8 @@ void listening_loop ()
     }
 }
 
-void sig_child(__attribute__((unused)) int signum)
-{ wwaitpid (-1,NULL,0); }
+void sig_child (__attribute__((unused)) int signum)
+{ wwaitpid (-1, NULL, 0); }
 
 int main (__attribute__((unused)) int argc, char *argv[])
 {
@@ -449,7 +450,7 @@ int main (__attribute__((unused)) int argc, char *argv[])
    * do not become zombies is to catch the SIGCHLD signal and perform
    * a wait(2) or similar. (https://man7.org/linux/man-pages/man2/sigaction.2.html)
    */
-  signal(SIGCHLD, sig_child);
+  signal (SIGCHLD, sig_child);
 
   fprintf (stderr, "[%ld] loading config...\n\n", (long) getpid ());
   char buf[BUFSIZ];
